+++ doc/Changelog.md
@@ -4,6 +4,8 @@
 
 # Release 0.6 (30-06-2016)
 
+* [179](https://github.com/HBPVIS/ZeroEQ/pull/179):
+  Fix http server blocking issues with libzmq 4.1.4
 * [170](https://github.com/HBPVIS/ZeroEQ/pull/170):
   Remove FlatBuffers dependency and support
 * [169](https://github.com/HBPVIS/ZeroEQ/pull/169):